Fix issue with Werkzeug and weird query string urldecoding
Werkzeug's request object has a url method that returns a URL with a
partially-decoded querystring, causing some methods to fail in weird ways with
a message about urlencoding.  This patch fixes the problem for me.

I have also added a couple tests for the method in the utils model that
extracts the params from the request object since it was untested.  The second
test is the regresssion test that reproduces the bug.

I have also removed a test that I think is incorrect.  See my comments on the
commit here:

https://github.com/lepture/flask-oauthlib/commit/7102484e423686e9bc31b32688531b5fa4e827e9