Change test for no content when parsing response
When parsing the response for a DELETE request to GitHub, I found I had
an empty byte-string instead of an empty string for my response data.
This caused OAuthlib to try and parse it with JSON, which will not work.
Changing the test to allow for any Falsey value will allow empty strings
or byte-strings to return an empty dict.