Remove incorrect handling of scope
The contrib client joins a list of scope with a comma which is wrong.
RFC6749 section 3.3 specifies the scopes must be "expressed as a list
of space-delimited, case-sensitive strings.".
Processing the scope at this point is also not necessary as the scope
are passed to requests-oauthlib and then oauthlib which can handle a
Python list just fine.